refactor(daemon): extract global middleware into middleware.ts
